home/django/agcs/requirements.txt
https://github.com/ryankaiser/python_agcs/archive/v3.0.2.tar.gz
anglerfish==1.0.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
css-html-js-minify==2.0.2
csscompressor==0.9.4
decorator==4.0.10
Django==1.9.6
django-assets==0.11
django-bootstrap3==7.0.1
django-htmlmin==0.9.1
django-minifiedstorage==0.3
django-recaptcha2==0.1.7
django-static==1.5.5
django-super-favicon==0.6.0
docutils==0.12
fabric-webbuilders==0.3
gitdb==0.6.4
GitPython==2.0.6
gunicorn==19.5.0
html5lib==0.9999999
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.3.0
ply==3.8
psycopg2==2.6.1
ptyprocess==0.5.1
python3-memcached==1.51
recaptcha-client==1.0.6
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
slimit==0.8.1
smmap==0.9.0
traitlets==4.2.1
webassets==0.11
